LibraryAPI/src/pyproject.toml
[project]
name = "LibraryAPI"
version = "0.1.0"
description = "Серверная часть приложения по управлению библиотекой"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.8",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
